More tweaks.
git-svn-id: http://svn.eol.ucar.edu/svn/eol/common/trunk/pentek@6843 da89c675-a4d7-0310-8a15-9e7546503515

+++ p7142.h
@@ -81,7 +81,7 @@ namespace Pentek {
 			/// Write the baseband signal to ram.
 			/// @param buf The I and Q signal values
 			/// @param n The number of values.
-			virtual void write(short* data, int n);
+			virtual void write(long* data, int n);
 			/// Start the DAC output
 			virtual void startDAC();
 			/// Stop the DAC output
